+++ projects/journey-maps-client/src/lib/journey-maps-client.component.ts
@@ -139,6 +139,7 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
   // **************************************** BUTTON OPTIONS *****************************************/
 
   private defaultButtonControlOptions: ButtonControlOptions = {
+    showSmallButtons: false,
     levelSwitch: true,
     zoomControls: true,
     basemapSwitch: true,