+++ projects/journey-maps-client/src/lib/journey-maps-client.component.spec.ts
@@ -141,7 +141,7 @@ const setupFixtureAndComponent = (oneFingerPan: boolean = true) => {
   component = fixture.componentInstance;
   component.apiKey = 'apiKey';
   component.markerOptions.markers = markers;
-  component.controlOptions.oneFingerPan = oneFingerPan;
+  component.interactionOptions.oneFingerPan = oneFingerPan;
   fixture.detectChanges();
 };